requirements.txt
streamlit
langchain
requests
python-dotenv
google-api-python-client
arxiv
yt-dlp
